# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Akasha API",
    "description": "Personal knowledge retrieval.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/activity": {
      "get": {
        "tags": [
          "activity"
        ],
        "summary": "Your activity, newest first: files uploaded, renamed, tagged, opened and\ndeleted, searches, questions, collection changes, and the security log\n(sign-ins with address and browser, password and session changes, tokens,\nrate limits). Kept for `AKASHA_ACTIVITY_RETENTION_DAYS` (default 365).",
        "operationId": "list_activity",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Comma-separated categories (`files,search,chat,collections,security`); default all.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Comma-separated kinds (e.g. `auth.signed_in,auth.sign_in_failed`); default all.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file_id",
            "in": "query",
            "description": "Only events about this file.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "`next_cursor` from the previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1-100 (default 50).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "activity"
        ],
        "summary": "Clear your activity history (all of it, or some categories). The security\nlog is kept: it ages out with the retention period.",
        "operationId": "clear_activity",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Comma-separated categories to clear (default: everything except `security`,\nwhich is never cleared by hand).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityCleared"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with email and password.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in; session cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "End the current session.",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create an account and sign in.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Signed in; session cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityCategory": {
        "type": "string",
        "description": "Broad groups of events (the timeline's filters).",
        "enum": [
          "files",
          "search",
          "chat",
          "collections",
          "security"
        ]
      },
      "ActivityCleared": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ActivityItem": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "category",
          "created_at",
          "details"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          },
          "collection_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The collection, while it still exists."
          },
          "conversation_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The conversation, while it still exists."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "details": {
            "type": "object",
            "description": "Kind-specific facts, e.g. `{\"from\": \"a.pdf\", \"to\": \"b.pdf\"}` for a rename,\n`{\"count\": 3, \"file_names\": [..]}` for collection changes, `{\"added\": [..],\n\"removed\": [..]}` for tags, `{\"results\": 4, \"mode\": \"hybrid\"}` for searches."
          },
          "file_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The file, while it still exists."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ip": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client address (security events only)."
          },
          "kind": {
            "$ref": "#/components/schemas/ActivityKind"
          },
          "subject": {
            "type": [
              "string",
              "null"
            ],
            "description": "What it was about, as named at the time: file name, search query,\nquestion, collection or token name."
          },
          "user_agent": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client user agent (security events only)."
          },
          "via": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ActivityVia"
              }
            ]
          }
        }
      },
      "ActivityKind": {
        "type": "string",
        "description": "What happened.",
        "enum": [
          "FileUploaded",
          "FileRenamed",
          "FileTagged",
          "FileDeleted",
          "FileOpened",
          "SearchPerformed",
          "ChatAsked",
          "CollectionCreated",
          "CollectionUpdated",
          "CollectionDeleted",
          "CollectionFilesAdded",
          "CollectionFilesRemoved",
          "AccountCreated",
          "SignedIn",
          "SignInFailed",
          "SignedOut",
          "PasswordChanged",
          "PasswordChangeFailed",
          "AccountDeleteFailed",
          "SessionRevoked",
          "TokenCreated",
          "TokenRevoked",
          "RateLimited"
        ]
      },
      "ActivityPage": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityItem"
            },
            "description": "Newest first."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass as `cursor` for older events; `null` on the last page."
          }
        }
      },
      "ActivityVia": {
        "type": "string",
        "description": "Who acted: the browser session or an API token / MCP client.",
        "enum": [
          "session",
          "token"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "created_at",
          "record_search_history"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "record_search_history": {
            "type": "boolean",
            "description": "Searches (with their query) appear in the activity timeline."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal API token (akasha_pat_…)"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "akasha_session"
      }
    }
  }
}